refactor(perf): rename business/ -> ops/
