+++ tests/perf/ops/cleanup.py
